# travel-agent-python/app/api/business/itinerary.py
"""行程业务域：已迁移的端点集中在此（对应 Java ItineraryController）。"""

from __future__ import annotations

from collections.abc import AsyncIterator
from typing import Any

from fastapi import APIRouter, Body, Depends, File, Header, Path, Query, UploadFile
from fastapi.responses import StreamingResponse
from starlette.concurrency import run_in_threadpool

from app.api.deps import AuthUser
from app.api.security import enforce_business_auth
from app.common import event_hub, event_publisher
from app.common.addons import require_addon
from app.common.config import settings
from app.common.envelope import ApiError, ok
from app.schemas.business.itinerary import (
    ApplyPlansBody,
    ArchiveBody,
    ChatEditBody,
    CoverBody,
    FavoriteBody,
    GenerateTripRequest,
    HotelOptionRequest,
    ItemUpsertRequest,
    NlEditBody,
    OptimizeDayBody,
    PreferenceSignalsBody,
    ShareCreateBody,
    UpdateDayBody,
    VersionBody,
)
from app.services import (
    cover_service,
    day_persistence,
    itinerary_chat,
    itinerary_city,
    itinerary_command,
    itinerary_generation,
    itinerary_nl_edit,
    itinerary_plan_apply,
    itinerary_query,
    itinerary_version,
    preferences,
    quota_service,
    share_service,
)

router = APIRouter(
    prefix="/api/itinerary",
    tags=["itinerary"],
    dependencies=[Depends(enforce_business_auth)],
)


@router.get("/supported-cities")
def get_supported_cities() -> dict:
    return ok(itinerary_city.supported_cities())


# ---- agent 能力（同进程直调，迁移前是 Java→HTTP 一跳）----


@router.post("/clarify")
def post_clarify(body: dict[str, Any], user: AuthUser = Depends(enforce_business_auth)) -> dict:
    """槽位澄清：纯解析，不落库（同 Java `ItineraryCityService.clarify`）。"""
    quota_service.enforce_llm_budget(user.id)
    slots = body.get("slots")
    return ok(itinerary_city.clarify(str(body.get("message") or ""), slots if isinstance(slots, dict) else {}))


@router.post("/city-guide")
def post_city_guide(body: dict[str, Any], user: AuthUser = Depends(enforce_business_auth)) -> dict:
    quota_service.enforce_llm_budget(user.id)
    history = body.get("history")
    return ok(itinerary_city.city_guide(str(body.get("input") or ""), history if isinstance(history, list) else []))


@router.post("/poi-nearby")
def post_poi_nearby(body: dict[str, Any], user: AuthUser = Depends(enforce_business_auth)) -> dict:
    quota_service.enforce_llm_budget(user.id)
    return ok(itinerary_city.poi_nearby(body))


@router.post("/generate")
def post_generate(
    body: GenerateTripRequest,
    user: AuthUser = Depends(enforce_business_auth),
    x_idempotency_key: str | None = Header(default=None, alias="X-Idempotency-Key"),
) -> dict:
    """建壳 + 异步逐日生成，立即返回可轮询的初始详情（status=1）。

    `X-Idempotency-Key`（可选）：网络层重试 / 双击时带同一个键，TTL 内返回
    同一个行程而不是重复建壳（backlog「被重复请求咬过」）。
    """
    quota_service.enforce_llm_budget(user.id)
    return ok(itinerary_generation.generate(user.id, body, idempotency_key=x_idempotency_key))


# ---- 偏好：字面量路径必须声明在 /{id} 之前，否则会被路径参数吞掉 ----


@router.get("/preferences")
def get_preferences(user: AuthUser | None = Depends(enforce_business_auth)) -> dict:
    return ok(top_preferences(user.id))


@router.post("/preferences/signals")
def post_preference_signals(
    body: PreferenceSignalsBody,
    user: AuthUser | None = Depends(enforce_business_auth),
) -> dict:
    preferences.record_signals(
        user.id,
        body.explicitPreferences,
        body.hardConstraints,
        body.negativePreferences,
        body.source,
        1.0 if body.confidence is None else body.confidence,
    )
    return ok()


@router.get("/preferences/signals")
def get_preference_signals(user: AuthUser | None = Depends(enforce_business_auth)) -> dict:
    return ok(preferences.signals(user.id, 50))


@router.get("")
def list_itineraries(
    view: str | None = Query(None, description="all|active|done|favorite|archived"),
    q: str | None = Query(None, description="按标题或城市模糊过滤"),
    user: AuthUser | None = Depends(enforce_business_auth),
) -> dict:
    return ok(itinerary_query.list_summaries(user.id, view, q))


@router.get("/{id}")
def get_itinerary(
    id: int = Path(..., ge=1),
    user: AuthUser | None = Depends(enforce_business_auth),
) -> dict:
    return ok(itinerary_query.detail(user.id, id))


def top_preferences(user_id: int, limit: int = 5) -> list[str]:
    """高频正向偏好（Java 侧固定取 top 5）。"""
    return preferences.top_preferences(user_id, limit)


# ---- 写路径与版本（M4）----


@router.post("/{id}/items")
def add_item(
    request: ItemUpsertRequest,
    id: int = Path(..., ge=1),
    user: AuthUser | None = Depends(enforce_business_auth),
) -> dict:
    return ok(itinerary_command.add_item(user.id, id, request))


@router.put("/items/{itemId}")
def update_item(
    request: ItemUpsertRequest,
    itemId: int = Path(..., ge=1),
    user: AuthUser | None = Depends(enforce_business_auth),
) -> dict:
    return ok(itinerary_command.update_item(user.id, itemId, request))


@router.post("/{id}/optimize")
def post_optimize(
    body: OptimizeDayBody,
    id: int = Path(..., ge=1),
    user: AuthUser | None = Depends(enforce_business_auth),
    _addon: None = Depends(require_addon("schedule_optimizer")),
) -> dict:
    """按路线重排某天（v2.6 W3）：确定性优化器，全量重排、不做锁定项、不删除点位。

    G-3.1：schedule_optimizer addon 停用时本端点 404（require_addon，隐藏而非 403）。
    """
    if body.dayId is None:
        raise ApiError(400, "dayId 必填")
    return ok(itinerary_command.optimize_day(user.id, id, body.dayId))


@router.patch("/{id}/days/{dayId}")
def patch_day(
    body: UpdateDayBody,
    id: int = Path(..., ge=1),
    dayId: int = Path(..., ge=1),
    user: AuthUser | None = Depends(enforce_business_auth),
) -> dict:
    """编辑日副标题（复用 metadata_json.theme；空串 = 清空回退派生标题）。"""
    return ok(itinerary_command.update_day(user.id, id, dayId, body.theme))


@router.post("/{id}/nl-edit")
def post_nl_edit(
    body: NlEditBody,
    id: int = Path(..., ge=1),
    user: AuthUser = Depends(enforce_business_auth),
) -> dict:
    """自然语言编辑：解析并**直接落库**（chat-edit 只出草稿，两者职责不同）。"""
    quota_service.enforce_llm_budget(user.id)
    return ok(itinerary_nl_edit.nl_edit(user.id, id, body.instruction))


@router.post("/{id}/apply-plans")
def post_apply_plans(
    body: ApplyPlansBody,
    id: int = Path(..., ge=1),
    user: AuthUser | None = Depends(enforce_business_auth),
) -> dict:
    return ok(itinerary_plan_apply.apply_plans(user.id, id, body.actionMessageId, body.baseRevision))


@router.post("/{id}/hotel-option")
def post_hotel_option(
    body: HotelOptionRequest,
    id: int = Path(..., ge=1),
    user: AuthUser | None = Depends(enforce_business_auth),
) -> dict:
    return ok(itinerary_plan_apply.apply_hotel_option(user.id, id, body))


@router.delete("/items/{itemId}")
def delete_item(itemId: int = Path(..., ge=1), user: AuthUser | None = Depends(enforce_business_auth)) -> dict:
    return ok(itinerary_command.delete_item(user.id, itemId))


@router.put("/{id}/days/{dayId}/order")
def reorder_items(
    # 请求体是裸 JSON 数组（Java `@RequestBody List<Long>`），不是 {"itemIds": [...]}
    item_ids: list[int],
    id: int = Path(..., ge=1),
    dayId: int = Path(..., ge=1),
    user: AuthUser | None = Depends(enforce_business_auth),
) -> dict:
    return ok(itinerary_command.reorder_items(user.id, id, dayId, item_ids))


@router.delete("/{id}")
def delete_itinerary(id: int = Path(..., ge=1), user: AuthUser | None = Depends(enforce_business_auth)) -> dict:
    itinerary_command.delete_itinerary(user.id, id)
    return ok()


@router.get("/{id}/chat-history")
def chat_history(id: int = Path(..., ge=1), user: AuthUser | None = Depends(enforce_business_auth)) -> dict:
    return ok(itinerary_chat.chat_history(user.id, id))


@router.delete("/{id}/chat-history")
def clear_chat_history(id: int = Path(..., ge=1), user: AuthUser | None = Depends(enforce_business_auth)) -> dict:
    itinerary_chat.clear_history(user.id, id)
    return ok()


@router.get("/{id}/events")
async def events(
    id: int = Path(..., ge=1), user: AuthUser | None = Depends(enforce_business_auth)
) -> StreamingResponse:
    """订阅生成进度事件流（进程内 SSE；取代 Java 的「Redis pub/sub → SseEmitter」转发桥）。"""
    # 归属查询进线程池，登记订阅必须在事件循环里做（要捕获 loop 才能跨线程投递）
    main = await run_in_threadpool(itinerary_query.find_readable_main, user.id, id)
    subscription, _rejected = event_hub.subscribe(id, event_publisher.too_many_connections_envelope(id))
    snapshot = await run_in_threadpool(_terminal_snapshot, id, main)
    return _sse(_event_frames(id, subscription, snapshot))


def _terminal_snapshot(itinerary_id: int, main) -> str | None:
    """订阅时行程已经终态 → 补一帧收尾，别让这条连接永远只发心跳（R2-2）。

    总线只广播、不留历史（`event_hub` 无回放缓冲），所以"生成结束之后才连上"
    （EventSource 瞬时断线重连、或用户在结果页重新打开）的客户端永远等不到
    终态帧。终态判定读库里的 `gen_state`，不猜缓存。
    """
    gen_state = getattr(main, "gen_state", None)
    if gen_state == "FAILED":
        return event_publisher.error_envelope(itinerary_id, "GENERATION_FAILED", "生成未完成，可重新发起生成")
    if gen_state not in ("COMPLETED", "PARTIAL"):
        return None
    expected = int(getattr(main, "days", 0) or 0)
    missing = set(day_persistence.unfinished_day_nos(itinerary_id))
    return event_publisher.build_envelope(
        itinerary_id,
        "done",
        {
            "daysExpected": expected,
            "daysEmitted": [day_no for day_no in range(1, expected + 1) if day_no not in missing],
            "tripTheme": getattr(main, "trip_theme", None),
            "complete": gen_state == "COMPLETED",
            "message": "该行程已生成完成",
        },
    )


async def _event_frames(itinerary_id: int, subscription, snapshot: str | None = None):
    try:
        if snapshot is not None:
            yield f"data:{snapshot}\n\n"
            return
        while True:
            envelope = await subscription.take()
            if envelope is event_hub.CLOSED:
                return
            if envelope is None:
                # 空闲一个心跳周期才补心跳帧：连接保持活跃，但业务事件永远优先
                envelope = event_publisher.heartbeat_envelope(itinerary_id)
            yield f"data:{envelope}\n\n"
    finally:
        # 客户端断开/服务收尾都要摘除，否则注册表里留死连接
        event_hub.unsubscribe(subscription)


@router.post("/{id}/chat-edit")
def chat_edit(
    id: int = Path(..., ge=1), body: ChatEditBody = Body(...), user: AuthUser = Depends(enforce_business_auth)
) -> dict:
    quota_service.enforce_llm_budget(user.id)
    return ok(itinerary_chat.chat_edit(user.id, id, body.message, body.history))


@router.post("/{id}/chat-edit/stream")
async def chat_edit_stream(
    id: int = Path(..., ge=1), body: ChatEditBody = Body(...), user: AuthUser = Depends(enforce_business_auth)
) -> StreamingResponse:
    """对话编辑的 SSE 变体：与非阻塞版同参构造、同一条落库收尾路径。"""
    quota_service.enforce_llm_budget(user.id)
    await run_in_threadpool(itinerary_query.find_writable_main, user.id, id)
    return _sse(_sse_frames(itinerary_chat.chat_edit_stream(user.id, id, body.message, body.history)))


async def _sse_frames(envelopes) -> AsyncIterator[str]:
    async for envelope in envelopes:
        yield f"data:{envelope}\n\n"


def _sse(frames) -> StreamingResponse:
    return StreamingResponse(
        frames,
        media_type="text/event-stream",
        headers={
            "Cache-Control": "no-cache",
            # 反向代理默认会缓冲响应，SSE 会被攒成一坨；这个头让 nginx 对该连接关掉缓冲
            "X-Accel-Buffering": "no",
        },
    )


@router.get("/{id}/versions")
def list_versions(id: int = Path(..., ge=1), user: AuthUser | None = Depends(enforce_business_auth)) -> dict:
    return ok(itinerary_version.list_versions(user.id, id))


@router.get("/{id}/versions/diff")
def versions_diff(
    fromVersionId: int = Query(..., ge=1),
    toVersionId: int = Query(..., ge=1),
    id: int = Path(..., ge=1),
    user: AuthUser | None = Depends(enforce_business_auth),
) -> dict:
    return ok(itinerary_version.diff(user.id, id, fromVersionId, toVersionId))


@router.post("/{id}/versions")
def create_version(
    body: VersionBody,
    id: int = Path(..., ge=1),
    user: AuthUser | None = Depends(enforce_business_auth),
) -> dict:
    return ok(itinerary_version.create_snapshot(user.id, id, body.operation, body.summary))


@router.post("/{id}/versions/{versionId}/restore")
def restore_version(
    id: int = Path(..., ge=1),
    versionId: int = Path(..., ge=1),
    user: AuthUser | None = Depends(enforce_business_auth),
) -> dict:
    return ok(itinerary_version.restore(user.id, id, versionId))


# ---- 封面 / 收藏 / 归档（SPEC v2.3 §6.3/§6.5，S1）----


@router.post("/{id}/cover")
def set_cover(
    body: CoverBody,
    id: int = Path(..., ge=1),
    user: AuthUser | None = Depends(enforce_business_auth),
) -> dict:
    """封面三来源：unsplash=服务端 snapshot 落盘；default=四个 cover 列置 NULL。"""
    if body.source == "unsplash":
        return ok(cover_service.set_cover_unsplash(user.id, id, body.unsplashId or ""))
    if body.source == "default":
        return ok(cover_service.set_cover_default(user.id, id))
    raise ApiError(400, "source 仅支持 unsplash|default")


@router.post("/{id}/cover/upload")
def upload_cover(
    file: UploadFile = File(...),
    id: int = Path(..., ge=1),
    user: AuthUser | None = Depends(enforce_business_auth),
) -> dict:
    """用户上传（jpeg/png，≤5MB——S0-3 在服务层流式截断）；文件名一律不使用（uuid 落盘）。"""
    data = cover_service.read_upload_capped(file.file, settings.cover_upload_max_bytes)
    return ok(
        cover_service.set_cover_upload(user.id, id, filename=file.filename, content_type=file.content_type, data=data)
    )


@router.post("/{id}/favorite")
def set_favorite(
    body: FavoriteBody,
    id: int = Path(..., ge=1),
    user: AuthUser | None = Depends(enforce_business_auth),
) -> dict:
    return ok(itinerary_command.set_favorite(user.id, id, body.favorite))


@router.post("/{id}/archive")
def set_archived(
    body: ArchiveBody,
    id: int = Path(..., ge=1),
    user: AuthUser | None = Depends(enforce_business_auth),
) -> dict:
    return ok(itinerary_command.set_archived(user.id, id, body.archived))


# ---- 公开分享（SPEC v2.3 §6.6，S2）：owner 面三端点；匿名面在 share.py ----


@router.post("/{id}/share")
def create_share(
    body: ShareCreateBody,
    id: int = Path(..., ge=1),
    user: AuthUser | None = Depends(enforce_business_auth),
) -> dict:
    """创建/轮换分享链接（重复创建会让旧链接立即失效）。"""
    return ok(share_service.create_share(user.id, id, body.expireDays))


@router.get("/{id}/share")
def get_share(
    id: int = Path(..., ge=1),
    user: AuthUser | None = Depends(enforce_business_auth),
) -> dict:
    return ok(share_service.get_share(user.id, id))


@router.delete("/{id}/share")
def remove_share(
    id: int = Path(..., ge=1),
    user: AuthUser | None = Depends(enforce_business_auth),
) -> dict:
    share_service.remove_share(user.id, id)
    return ok()